requirements.txt
MouseInfo==0.1.3
packaging==21.3
Pillow==9.3.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytesseract==0.3.10
pytweening==1.0.4
